refactor(about): move ocean-watch dashboard `About` page up to top-level
